// examples/express-prisma/schema.generated.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface BaseRouter {
  nodeType: "router";
  routes: {
    user: BaseUserRouter;
  };
}
export interface BaseUserRouter {
  nodeType: "router";
  routes: {
    create: BaseUserCreateRoute;
    test: BaseUserTestRoute;
  };
}
export interface BaseUserCreateRoute {
  type: "mutation";
  nodeType: "route";
  schemaArgs: {
    name: string;
    email: string;
  };
  schemaOutput: {
    id: number;
    email: string;
    name: string | null;
  };
}
export interface BaseUserTestRoute {
  type: "query";
  nodeType: "route";
  schemaArgs: unknown;
  schemaOutput: string;
}